rfms snapshot fix

+++ dotnet/corecomponents/net.atos.daf.ct2.rfms/response/RfmsVehicleStatus.cs
@@ -185,7 +185,7 @@ namespace net.atos.daf.ct2.rfms.response
         public double? TachographSpeed { get; set; }
         public int? EngineSpeed { get; set; }
         public string FuelType { get; set; }
-        public int? FuelLevel1 { get; set; }
+        public double? FuelLevel1 { get; set; }
         //public int? FuelLevel2 { get; set; }
         public int? CatalystFuelLevel { get; set; }
         public string Driver1WorkingState { get; set; }
